Move router-lib impl out of index.ts
