+++ packages/microapps-router-lib/src/monolith.prefix.spec.ts
