+++ packages/microapps-router-lib/src/monolith.spec.ts
